size image to window
The default camera might not be sized at 640x480 and so a partial image will be shown in most cases (which I found has confused some users).  This fixes that issue though the image might also be skewed.  I'm not sure the best way to handle this whether in `draw()` or in `setup()`